move `telemetry` to `extension` as module
`telemetry` has been always an extension module, it makes more sense for it to live there
