+++ bottlecap/src/extension/telemetry/events.rs
