+++ bottlecap/src/lifecycle/invocation/processor.rs
@@ -13,6 +13,7 @@ use tracing::{debug, warn};
 
 use crate::{
     config::{self, aws::AwsConfig},
+    extension::telemetry::events::{InitType, ReportMetrics, RuntimeDoneMetrics, Status},
     lifecycle::invocation::{
         base64_to_string,
         context::{Context, ContextBuffer, ReparentingInfo},
@@ -25,7 +26,6 @@ use crate::{
         constants::{ETC_PATH, PROC_PATH},
     },
     tags::{lambda::tags::resolve_runtime_from_proc, provider},
-    telemetry::events::{InitType, ReportMetrics, RuntimeDoneMetrics, Status},
     traces::{
         context::SpanContext,
         propagation::{
