+++ bottlecap/tests/logs_integration_test.rs
@@ -1,9 +1,9 @@
 use bottlecap::LAMBDA_RUNTIME_SLUG;
 use bottlecap::config::Config;
 use bottlecap::event_bus::EventBus;
+use bottlecap::extension::telemetry::events::TelemetryEvent;
 use bottlecap::logs::{agent::LogsAgent, flusher::LogsFlusher};
 use bottlecap::tags::provider::Provider;
-use bottlecap::telemetry::events::TelemetryEvent;
 use dogstatsd::api_key::ApiKeyFactory;
 use httpmock::prelude::*;
 use std::collections::HashMap;
